Escape special chars (like +) in html attribute name.
In <a name=... HTML attributes the attribute value needs to begin with a
letter (A-Za-z) and then may be followed by any number of letters, digits
hyphens, underscores, colons and periods.

Other chars, like + or ~ are not allowed.  Now we encode these special
characters as :<hexvalue>:.  We do not do anything yet to make sure
these values start with a letter.