Unicode related argument fixes
python's urllib does not want to encode non-ASCII strings.  While we
do not really care how they are encoded, we very much would prefer
it not to throw an exception.

So we encode it somehow so we can do a sane redirect from
/package/?src=foobar to /package/foobar/.

Also, handle non-ASCII package names in /package/foobar/
before passing it on to the DB layer which likewise will
complain that it cannot handle such strings.  Non-ASCII
named packages do not exist anyway, so we can bail out with
404 error early.